Multi Criteria Search Functionality and Search Function Initializer
We added in the functionality to allow multiple criteria to be entered simultaneously on user input for searches. For this to function I also had to tweak some previous functionality to stream line the process, as well as fix some minor bugs. I was also able to maintain the ability to further narrow searches after this if the user wished.